Simplify SDKUtil by using Java8 Files.walkFileTree
This commit does not add any new feature, but re-implements
some functions in SDKUtil to use JDK8 features to traverse
the file tree.
Also moved the Docker files and logging properties to src/main/resources
as they are not included in the built jar file if placed inside the
java folder.

Change-Id: I32938e4c0f1c26663dec8848a76721fe25784810
Signed-off-by: Muhammad Altaf <muhammada@fast.au.fujitsu.com>