FAB-928 Endorser proposal update to V1.0
Definitely work in progress..
Jump start to Fabric V 1.00 sdk.
 The protobufs are up to date with about a day ago with master Fabric.
The README.md has the specific commit level of the Fabric and how to get to it.
At that level it can create a chain, deploy, modify, and query the ledger.
Some clean up is yet needed too.
There has been some refactoring to V1.0 spec mostly for the public facing classes.
One goal was to NOT expose any of the Fabric generated artifacts to insulate
the user from changes when possible.

Change-Id: I646bcee73a37db9c525093905f7599ee71bae7ff
Signed-off-by: rickr <cr22rc@gmail.com>