[FAB-1826]  Verify peer cert chain of trust
We get the peer cert when checking a ProposalResponse.
Validate the peer cert chain of trust by checking
against the SDK's trust store.

For now, we copy the peer's CA cert and add
it to the SDK trust store under the covers.

This will be cleaned up once the MSP work is
complete and the SDK trust store is populated
from the chain's genesis block.

As a side effect, there is now a properties file
that the sdk will automatically read in at init time.
File is hyperledgerfabricsdk.properties

Feb 2, 2017 rebase on master

Change-Id: Ia148bfc4677bd9abf88aec7222a0e7d57d42615d
Signed-off-by: tuand27613 <tdang@us.ibm.com>