[FAB-1826] [WIP] Verify peer cert chain of trust
We get the peer cert when checking a ProposalResponse.
Validate the peer cert chain of trust by checking
against the SDK's trust store.

For now, we copy the peer's CA cert and add
it to the SDK trust store under the covers.

This will be cleaned up once the MSP work is
complete and the SDK trust store is populate
from the chain's genesis block.

Change-Id: Ia148bfc4677bd9abf88aec7222a0e7d57d42615d
Signed-off-by: tuand27613 <tdang@us.ibm.com>