[FAB-6888] Reduce logging for Maven downloads
See the JIRA for an example of the logs, but
basically the CI logs are full of download and
status logs for pulling from Maven repos.

This change does two things:
- runs Maven in batchmode to eliminate the download
progress messages
- sets the logging level to warn for downloading
from Maven repos in general

Technically only the first is needed, but in case
one wants to locally see the downloads both settings
are used.

Change-Id: I3d511e2cd770e718fc8846da3b39f78cca90c256
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>