[FAB-6836] Implement Idemix Signing Identity
This commit adds the Identity and SigningIdentity interfaces
and their implementation with Idemix crypto library
for transaction signing and further integration with the sdk.
The tests use crypto material (configs) that is generated by
the idemixgen tool from fabric, so the crypto implementations
in Go and Java are fully compatible.

Change-Id: I13d43bac6e87be36fd46fa77b6762aa56b9b1e0d
Signed-off-by: Maria Dubovitskaya <mdu@zurich.ibm.com>
Signed-off-by: Manu Drijvers <mdr@zurich.ibm.com>
Signed-off-by: Saad Karim <skarim@us.ibm.com>