Signed-off-by: Alan Atkinson <alan@alanatkinson.net>
[FAB-9616] Add example to allow adding a member to a consortium and a channel

Sample code to show adding a third member to the consortium created in End2EndIT.
Additionally it adds the member to the foo channel, inatantiates a new version of the chaincode on all 3 orgs and runs a sample transaction through them.

Change-Id: Ide3b9d0e008e4307c3f7180e9707a53dd941a2b2

[FAB-9616] Provide sample updating channel with new organization

Example adds a new org to the consortium from End2EndIt.
It then adds the org to the foo channel, instantiates some chaincode and
runs a transaction through the three org peers to show it all works.

Signed-off-by: aatkddny <alan@alanatkinson.net>

[FAB-9616] Example of adding an org to a constortium and a channel

Adds a new org to the consortium established in End2EndIT.
Then adds the org to the "foo" channel and loads a new copy of chaincode on the peers for all three orgs.
It then runs a transaction through with an AND requirement for consensus against all three to show
functionality.

Signed-off-by: aatkddny <alan@alanatkinson.net>

[FAB-9616] Example of adding a new org

Adds a new org to the consortium
Adds the new org to the foo channel
Installs a new copy of the chaincode on all three org peers
Runs a transaction through them to show it all works

Change-Id:
Singed-off-by: aatkddny <alan@alanatkinson.net>

[FAB-9616] Example of adding a new org

Adds a new org to the consortium from End2EndIT.
Adds the org to the foo channel.
Loads new chaincode on the peers for all three orgs.
Runs a Tx through to be sure it all works.

Change-Id:
Signed-off-by: aatkddny <alan@alanatkinson.net>

[FAB-9616] Example of adding a new org

Adds a new org to the consortium from End2EndIT.
Adds the org to the foo channel.
Loads new chaincode on the peers for all three orgs.
Runs a Tx through to be sure it all works.

Change-Id:
Signed-off-by: aatkddny <alan@alanatkinson.net>

[FAB-9616] Add example to allow adding a member to a consortium and a channel

Sample code to show adding a third member to the consortium created in End2EndIT.
Additionally it adds the member to the foo channel, inatantiates a new version of the chaincode on all 3 orgs and runs a sample transaction through them.

Change-Id: Ide3b9d0e008e4307c3f7180e9707a53dd941a2b2
Signed-off-by: aatkddny <alan@alanatkinson.net>