[FAB-10230] Add self-signed TLS cert generation
This change set adds a utility to generate self-signed
client (and server) TLS certificates.

It is useful for users that don't have a TLS certificate,
but they want to use service discovery which requires mutual TLS.

When the peer is running in the mode which it doesn't verify the client
certificate, but still asks for it - the client can use a self-signed cert
in order to connect to the discovery service.

Change-Id: I734b47198b43c17daaa90dec711758ff035980e5
Signed-off-by: yacovm <yacovm@il.ibm.com>