[WIP] Bridge between Proposal and Transaction
This change-set allows to reuse a SigningIdentity,
created at TransactionContext construction and
used to generate a Proposal, at transaction generation
time. In the idemix context, a SigningIdentity represents
a pseudodym. A fresh pseudonym needs to be generated
everytime a new transaction cycle is initiated.
Therefore, we make sure the same psuedonym is used
at proposal and transaction generation time.

Change-Id: I28cc0d0232c83abd19471b1a2cd39f8735c08015
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>