[FABJ-xxxxx] Add Go Modules to Go Chaincode
Added Go modules to the Go chaincode contained in
the SDK Integration tests. Because of the method
used to package chaincode in the Java SDK the
vendor directory was also included until the SDK
can support the new Go Module chaincode packaging
model. The go.mod files pin the dependencies
for future implementations the support the new
model to refer to.

Signed-off-by: Brett Logan <Brett.T.Logan@ibm.com>
Change-Id: If0749554968c1f4be5ae82bd45a750d2f6d50dab