FGJ-51: Disable IdeMix; increase proposal timeout
IdeMix tests seem to be consistently failing in the master branch
only when nothing has changed in the client-side IdeMix code.
Temporarily disabling these tests to get build stability back
while the issue is investigated.

The response for the first proposal arriving at a given chaincode
on a given peer can take significantly longer as the chaincode
container needs to be built and started. With the previous default
timeout, intermittent failures were observed in automated scenario
testing.

Change-Id: I38491f2d9634720583968f833d0ddd2defc29ff1
Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>